style(cpc,docs,vendor): drop redundant literal casts, float suffixes and variant type args
The exemplars first, since they teach the style: `cpc init`'s templates and
docs/lang/skill.md, which cpc embeds and every agent is told to read before
writing C+. Then twelve packages, each gated on its own suite.

WHAT WAS NEVER REQUIRED

Verified by compiling both spellings, not by assumption:

  (0 as usize)          ->  0        in argument, annotated-let, struct-field,
                                     index, comparison, return and array-repeat
                                     position, including into u8, usize and i64
  0.0f64                ->  0.0
  Option[Text]::Some(v) ->  Option::Some(v)   in match / if let / while let /
                                              guard let — construction still
                                              needs the args (E0303)
  { *p } / { call(); }  ->  *p / call();      there is no unsafe block in this
                                              language; skill.md:666 says so

skill.md stated the variant rule at line 398 and then broke it in seven of its
own examples. It also wrote `let s: str = { owner.view() };` in the view-lifetime
section — the braces do nothing there, the rule is that the OWNER is named, and
that is the example a reader copies most carefully.

12 packages, 31 files, 4,352 tokens removed. All suites green: agent_core 228,
agent_mcp 361, agent_inapp 227, agent_appkit 325, facet_agent 560, inspector
701, json 179, flex_layout 309, events 72, log 15, uuid 159, arena 16 —
unchanged counts, so nothing was quietly dropped.

GENERATED FILES ARE SKIPPED

106 of them carry a generation banner. A cast removed there survives until the
next regen; the generator is where it would have to change.

WHAT STOPPED THE REST — bugs/int-literal-in-an-i64-slot-is-stored-as-i32.md

`stdlib`, `facet` and `facet_runtime` are REVERTED, because the substitution
reaches a pre-existing codegen defect the casts were hiding: an integer literal
type-checks against `i64` and is then emitted as `i32` into the 64-bit slot.

  enum V { Int(i64) }
  get(V::Int(10 as i64))  ==  get(V::Int(10))     is FALSE
  0x…0000_000a                0x…0001_0000_000a

Both print 10 truncated to i32. In stdlib the same defect is invalid IR that
clang rejects outright — `'%t20' defined with type 'i32' but expected 'i64'` —
which `cpc check` could never catch, since it does not run clang.

Position-specific, not general: annotated let, function argument, struct field,
`u64` and `f64` payloads and a two-field payload all widen correctly; a single
`i64` enum payload does not. Nobody had hit it because nobody writes the concise
form the type checker already accepts.

The cleanup tool now refuses to touch anything inside a `::Variant(...)` call,
so a cast in an enum payload stays load-bearing until this is fixed.

NOT DONE

Trailing `return;` in unit functions — 12,758 sites — is left alone. It is not
required by the compiler, but CLAUDE.md, skill.md and the style guide all
mandate it, so removing it is a decision rather than a cleanup.

+++ cpc/src/main.rs
@@ -6916,11 +6916,11 @@ fn run_init(args: &[OsString]) -> ExitCode {
          import \"stdlib/vec\" as vec;\n\n\
          struct Home {{\n    taps: i64,\n}}\n\n\
          impl Home {{\n\
-         \x20   fn new() -> Home {{ return Home {{ taps: 0 as i64 }}; }}\n\n\
+         \x20   fn new() -> Home {{ return Home {{ taps: 0 }}; }}\n\n\
          \x20   // A handler is a bound METHOD — `on_click: this.on_tap` wires it, and\n\
          \x20   // the compiler fills the context slot. Never hand-roll `#addr_of(this)`.\n\
          \x20   fn on_tap(ref this, sender: *u8) {{\n\
-         \x20       this.taps = this.taps + (1 as i64);\n\
+         \x20       this.taps = this.taps + 1;\n\
          \x20       this.show_taps();\n\
          \x20       return;\n\
          \x20   }}\n\n\
@@ -6939,14 +6939,14 @@ fn run_init(args: &[OsString]) -> ExitCode {
          \x20       return @ui {{\n\
          \x20           column {{\n\
          \x20               label(\"Hello from C+\", key: \"hello\",\n\
-         \x20                     font_size: 28.0f64,\n\
+         \x20                     font_size: 28.0,\n\
          \x20                     font_weight: vocab::FontWeight::Bold)\n\
          \x20               label(\"tapped 0\", key: \"taps\")\n\
          \x20               button(\"Tap me\", key: \"tap\", on_click: this.on_tap)\n\
          \x20           }}\n\
-         \x20               .grow(1.0f64)\n\
-         \x20               .gap(12.0f64)\n\
-         \x20               .padding(24.0f64)\n\
+         \x20               .grow(1.0)\n\
+         \x20               .gap(12.0)\n\
+         \x20               .padding(24.0)\n\
          \x20               .align(flex::Align::Center)\n\
          \x20               .justify(flex::Justify::Center)\n\
          \x20       }};\n\
@@ -6962,7 +6962,7 @@ fn run_init(args: &[OsString]) -> ExitCode {
          \x20       // window — but they are the same facade on both platforms and the\n\
          \x20       // iOS backend drops them.\n\
          \x20       return screen::Chrome::new(title: \"{proj_name}\",\n\
-         \x20                                  width: 390.0f64, height: 844.0f64);\n\
+         \x20                                  width: 390.0, height: 844.0);\n\
          \x20   }}\n\
          \x20   fn menu_items(this) -> vec::Vec[screen::MenuItem] {{\n\
          \x20       return vec::new::[screen::MenuItem]();\n\
@@ -7001,7 +7001,7 @@ fn run_init(args: &[OsString]) -> ExitCode {
          \x20   inspect::arm();\n\
          \x20   runtime::agent_mcp(\"{proj_name}\");\n\
          \x20   runtime::run_screen(Home::new());\n\
-         \x20   return 0 as i32;\n\
+         \x20   return 0;\n\
          }}\n"
     );
 
@@ -7467,7 +7467,7 @@ fn run_init(args: &[OsString]) -> ExitCode {
          import \"facet/services\" as services;\n\
          import \"stdlib/text\" as text;\n\n\
          fn answered(index: i32, ctx: *u8) {{\n\
-         \x20   if index == (0 as i32) {{ consent::allow_pending(); return; }}\n\
+         \x20   if index == 0 {{ consent::allow_pending(); return; }}\n\
          \x20   consent::deny_pending();\n\
          \x20   return;\n\
          }}\n\n\
